Added negative unit tests

+++ practice_app/views.py
@@ -2,9 +2,12 @@ import csv
 from io import TextIOWrapper
 
 from django.core.exceptions import ValidationError
-from django.db import IntegrityError
+from django.db import IntegrityError, transaction
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.urls import reverse_lazy
+from django.utils.decorators import method_decorator
+from django.views.decorators.http import require_http_methods
 from django.views.generic import UpdateView, CreateView
 
 from practice_app.forms import UploadCSVFileForm, EditOrCreateCSVRowForm
@@ -12,7 +15,8 @@ from practice_app.models import MuseumAPICSV
 from django.shortcuts import get_object_or_404
 
 
-def index(request):
+@require_http_methods(["GET"])
+def list_csv_content_view(request):
     """
     Homepage of the website. Displays uploaded CSV file in table form.
     :param request: HTTPRequest object
@@ -39,6 +43,10 @@ class CSVEditRowView(UpdateView):
     success_url = reverse_lazy('practice_app:index')
     required_css_class = 'required'
 
+    @method_decorator(require_http_methods(["GET", "POST"]))
+    def dispatch(self, *args, **kwargs):
+        return super().dispatch(*args, **kwargs)
+
 
 class CSVAddNewRowView(CreateView):
     """
@@ -50,21 +58,30 @@ class CSVAddNewRowView(CreateView):
     success_url = reverse_lazy('practice_app:index')
     required_css_class = 'required'
 
+    @method_decorator(require_http_methods(["GET", "POST"]))
+    def dispatch(self, *args, **kwargs):
+        return super().dispatch(*args, **kwargs)
 
+
+@require_http_methods(["GET", "POST"])
 def csv_row_delete_view(request, objectId):
     """
-        View that allows the user to upload CSV file to the server.
-        :param request: HTTPRequest object
-        :param objectId: id of the row to be deleted.
-        :return: HttpResponse object
-        """
+    View that allows the user to upload CSV file to the server.
+    :param request: HTTPRequest object
+    :param objectId: id of the row to be deleted.
+    :return: HttpResponse object
+    """
     if request.method == 'GET':
+        return HttpResponse(status=200)
+
+    if request.method == 'POST':
         row_obj = get_object_or_404(MuseumAPICSV, pk=objectId)
 
         row_obj.delete()
         return redirect('practice_app:index')
 
 
+@require_http_methods(["GET", "POST"])
 def csv_file_upload_view(request):
     """
     View that allows the user to upload CSV file to the server.
@@ -100,71 +117,72 @@ def csv_file_upload_view(request):
             for column in row:
                 columns.append(column)
             try:
-                MuseumAPICSV.objects.create(
-                    pk=columns[0],
-                    isHighlight=columns[1],
-                    accessionNumber=columns[2],
-                    accessionYear=columns[3],
-                    isPublicDomain=columns[4],
-                    primaryImage=columns[5],
-                    primaryImageSmall=columns[6],
-                    additionalImages=columns[7] or '',
-                    department=columns[8] or '',
-                    objectName=columns[9] or '',
-                    title=columns[10] or '',
-                    culture=columns[11] or '',
-                    period=columns[12] or '',
-                    dynasty=columns[13] or '',
-                    reign=columns[14] or '',
-                    portfolio=columns[15] or '',
-                    artistRole=columns[16] or '',
-                    artistPrefix=columns[17] or '',
-                    artistDisplayName=columns[18] or '',
-                    artistDisplayBio=columns[19] or '',
-                    artistSuffix=columns[20] or '',
-                    artistAlphaSort=columns[21] or '',
-                    artistNationality=columns[22] or '',
-                    artistBeginDate=columns[23] or '',
-                    artistEndDate=columns[24] or '',
-                    artistGender=columns[25] or 'm',
-                    artistWikidata_URL=columns[26] or '',
-                    artistULAN_URL=columns[27] or '',
-                    objectDate=columns[28] or '',
-                    objectBeginDate=columns[29] or '',
-                    objectEndDate=columns[30] or '',
-                    medium=columns[31] or '',
-                    dimensions=columns[32] or '',
-                    measurements=columns[33] or '',
-                    creditLine=columns[34] or '',
-                    geographyType=columns[35] or '',
-                    city=columns[36] or '',
-                    state=columns[37] or '',
-                    county=columns[38] or '',
-                    country=columns[39] or '',
-                    region=columns[40] or '',
-                    subregion=columns[41] or '',
-                    locale=columns[42] or '',
-                    locus=columns[43] or '',
-                    excavation=columns[44] or '',
-                    river=columns[45] or '',
-                    classification=columns[46] or '',
-                    rightsAndReproduction=columns[47],
-                    linkResource=columns[48] or '',
-                    metadataDate=columns[49] or '',
-                    repository=columns[50] or '',
-                    objectURL=columns[51] or '',
-                    tags=columns[52] or '',
-                    objectWikidata_URL=columns[53] or '',
-                    isTimelineWork=columns[54] or '',
-                    galleryNumber=columns[55] or -1,
-                    constituentID=columns[56] or -1,
-                    role=columns[57] or '',
-                    name=columns[58] or '',
-                    constituentULAN_URL=columns[59] or '',
-                    constituentWikidata_URL=columns[60] or '',
-                    gender=columns[61] or 'male',
+                with transaction.atomic():
+                    MuseumAPICSV.objects.create(
+                        pk=columns[0],
+                        isHighlight=columns[1],
+                        accessionNumber=columns[2],
+                        accessionYear=columns[3],
+                        isPublicDomain=columns[4],
+                        primaryImage=columns[5],
+                        primaryImageSmall=columns[6],
+                        additionalImages=columns[7] or '',
+                        department=columns[8] or '',
+                        objectName=columns[9] or '',
+                        title=columns[10] or '',
+                        culture=columns[11] or '',
+                        period=columns[12] or '',
+                        dynasty=columns[13] or '',
+                        reign=columns[14] or '',
+                        portfolio=columns[15] or '',
+                        artistRole=columns[16] or '',
+                        artistPrefix=columns[17] or '',
+                        artistDisplayName=columns[18] or '',
+                        artistDisplayBio=columns[19] or '',
+                        artistSuffix=columns[20] or '',
+                        artistAlphaSort=columns[21] or '',
+                        artistNationality=columns[22] or '',
+                        artistBeginDate=columns[23] or '',
+                        artistEndDate=columns[24] or '',
+                        artistGender=columns[25] or 'm',
+                        artistWikidata_URL=columns[26] or '',
+                        artistULAN_URL=columns[27] or '',
+                        objectDate=columns[28] or '',
+                        objectBeginDate=columns[29] or '',
+                        objectEndDate=columns[30] or '',
+                        medium=columns[31] or '',
+                        dimensions=columns[32] or '',
+                        measurements=columns[33] or '',
+                        creditLine=columns[34] or '',
+                        geographyType=columns[35] or '',
+                        city=columns[36] or '',
+                        state=columns[37] or '',
+                        county=columns[38] or '',
+                        country=columns[39] or '',
+                        region=columns[40] or '',
+                        subregion=columns[41] or '',
+                        locale=columns[42] or '',
+                        locus=columns[43] or '',
+                        excavation=columns[44] or '',
+                        river=columns[45] or '',
+                        classification=columns[46] or '',
+                        rightsAndReproduction=columns[47],
+                        linkResource=columns[48] or '',
+                        metadataDate=columns[49] or '',
+                        repository=columns[50] or '',
+                        objectURL=columns[51] or '',
+                        tags=columns[52] or '',
+                        objectWikidata_URL=columns[53] or '',
+                        isTimelineWork=columns[54] or '',
+                        galleryNumber=columns[55] or -1,
+                        constituentID=columns[56] or -1,
+                        role=columns[57] or '',
+                        name=columns[58] or '',
+                        constituentULAN_URL=columns[59] or '',
+                        constituentWikidata_URL=columns[60] or '',
+                        gender=columns[61] or 'male',
 
-                )
+                    )
             except IntegrityError:
                 pass
